# lib/api-zod/src/generated/types/createMenuItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Max Foods API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuItemRequest {
  restaurantId: string;
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  category?: string;
  isAvailable?: boolean;
}