# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Max Foods API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderItemRequest } from "./createOrderItemRequest";

export interface CreateOrderRequest {
  userId: string;
  restaurantId: string;
  deliveryAddress: string;
  deliveryFee: number;
  items: CreateOrderItemRequest[];
}