# lib/api-zod/src/generated/types/restaurant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Max Foods API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Restaurant {
  id: string;
  name: string;
  description?: string;
  category: string;
  imageUrl?: string;
  rating: number;
  deliveryTime: number;
  deliveryFee: number;
  minOrder: number;
  isActive: boolean;
  createdAt: string;
}